rename opre_ops to ops_api
