+++ .secrets.baseline
@@ -87,7 +87,7 @@
         "verified_result": null
       }
     ],
-    "backend/opre_ops/django_config/settings/local.py": [
+    "backend/ops_api/django_config/settings/local.py": [
       {
         "hashed_secret": "60ca8b161ee50e40662c3664e2701456e7eae82b",
         "is_secret": true,
@@ -97,7 +97,7 @@
         "verified_result": null
       }
     ],
-    "backend/opre_ops/ops_site/fixtures/fake_data.json": [
+    "backend/ops_api/ops_site/fixtures/fake_data.json": [
       {
         "hashed_secret": "6b8a1c79a05009129a1f92df6b370300adbabd56",
         "is_secret": false,
