+++ backend/.coveragerc
@@ -1,7 +1,7 @@
 [run]
 omit =
-        opre_ops/ops_site/migrations/*
-        opre_ops/ops_site/admin.py
+        ops_api/ops_site/migrations/*
+        ops_api/ops_site/admin.py
         **/tests/**
         *__init__*
