+++ backend/ops_api/django_config/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for opre_ops project.
+ASGI config for ops_api project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,7 +11,7 @@ import os
 from django.core.asgi import get_asgi_application
 
 os.environ.setdefault(
-    "DJANGO_SETTINGS_MODULE", "opre_ops.django_config.settings.cloudgov"
+    "DJANGO_SETTINGS_MODULE", "ops_api.django_config.settings.cloudgov"
 )
 
 application = get_asgi_application()
