+++ backend/ops_api/django_config/settings/common.py
@@ -8,7 +8,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # Application definition
 
 INSTALLED_APPS = [
-    "opre_ops.django_config.apps.OpsSiteConfig",
+    "ops_api.django_config.apps.OpsSiteConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
@@ -31,7 +31,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "opre_ops.django_config.urls"
+ROOT_URLCONF = "ops_api.django_config.urls"
 
 # TEMPLATES is needed for the admin application to run
 TEMPLATES = [
@@ -50,7 +50,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "opre_ops.django_config.wsgi.application"
+WSGI_APPLICATION = "ops_api.django_config.wsgi.application"
 
 
 # Password validation
