+++ backend/ops_api/django_config/settings/local.py
@@ -5,8 +5,8 @@ See the project Dockerfile and docker-compose.yml for context.
 These are settings for local development only, not cloud or production environments.
 """
 # Import all common settings relevant to both local & cloud:
-from opre_ops.django_config.settings.common import *  # noqa: F403, F401
-from opre_ops.django_config.settings.helpers.random_string import generate_random_string
+from ops_api.django_config.settings.common import *  # noqa: F403, F401
+from ops_api.django_config.settings.helpers.random_string import generate_random_string
 
 
 # Local database config for use with Docker Compose
