+++ backend/ops_api/django_config/settings/test.py
@@ -2,7 +2,7 @@
 Configuration for running OPRE OPS in a test
 """
 # Import all settings from the local environment
-from opre_ops.django_config.settings.local import *  # noqa: F403, F401
+from ops_api.django_config.settings.local import *  # noqa: F403, F401
 
 
 DATABASES = {
