+++ backend/ops_api/ops_site/admin.py
@@ -1,12 +1,12 @@
 from django.contrib import admin
 
-from opre_ops.ops_site.cans.models import CAN
-from opre_ops.ops_site.cans.models import CANFiscalYear
-from opre_ops.ops_site.cans.models import Contract
-from opre_ops.ops_site.cans.models import ContractLineItem
-from opre_ops.ops_site.cans.models import FundingPartner
-from opre_ops.ops_site.models import Person
-from opre_ops.ops_site.models import Role
+from ops_api.ops_site.cans.models import CAN
+from ops_api.ops_site.cans.models import CANFiscalYear
+from ops_api.ops_site.cans.models import Contract
+from ops_api.ops_site.cans.models import ContractLineItem
+from ops_api.ops_site.cans.models import FundingPartner
+from ops_api.ops_site.models import Person
+from ops_api.ops_site.models import Role
 
 
 @admin.register(FundingPartner)
