+++ backend/ops_api/ops_site/urls.py
@@ -1,10 +1,10 @@
 from django.urls import path
 
-from opre_ops.ops_site.cans.controller import CANFiscalYearByCanListController
-from opre_ops.ops_site.cans.controller import CanListController
-from opre_ops.ops_site.cans.controller import CanReadController
-from opre_ops.ops_site.portfolios.controller import PortfolioListController
-from opre_ops.ops_site.portfolios.controller import PortfolioReadController
+from ops_api.ops_site.cans.controller import CANFiscalYearByCanListController
+from ops_api.ops_site.cans.controller import CanListController
+from ops_api.ops_site.cans.controller import CanReadController
+from ops_api.ops_site.portfolios.controller import PortfolioListController
+from ops_api.ops_site.portfolios.controller import PortfolioReadController
 
 
 urlpatterns = [
