+++ backend/pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
 pythonpath = .
-DJANGO_SETTINGS_MODULE = opre_ops.django_config.settings.test
+DJANGO_SETTINGS_MODULE = ops_api.django_config.settings.test
 testpaths = tests
 addopts = --cov
